Tutorials: Changement de noms et correction de certaines scènes
