+++ tutorials/getting_started/tutorial_01_discretization_and_visualization.py
