+++ tutorials/getting_started/tutorial_02_gravity_and_dynamics.py
@@ -18,26 +18,22 @@ Key improvements over manual approach:
 import os
 import sys
 
-_beam_radius = 0.5
-_beam_length = 30.
-_nb_section = 32
-force_null = [0., 0., 0., 0., 0., 0.]  # N
-
 # Add the python package to the path
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", "..", "python"))
 
-from python.cosserat import BeamGeometryParameters, CosseratGeometry
+from cosserat import BeamGeometryParameters, CosseratGeometry
 
-from force_controller import ForceController
 from introduction_and_setup import (_add_cosserat_frame, _add_cosserat_state,
                                     _add_rigid_base, add_mini_header)
 
-v_damping_param: float = 8.e-1  # Damping parameter for dynamics
+v_damping_param: float =  8.e-1  # Damping parameter for dynamics
 
 def createScene(root_node):
     """Create a Cosserat beam scene with forces and dynamics."""
     # Configure scene with time integration
     add_mini_header(root_node)
+    root_node.addObject('RequiredPlugin', name='Sofa.Component.LinearSolver.Direct') # Needed to use components [SparseLDLSolver]  
+    root_node.addObject('RequiredPlugin', name='Sofa.Component.ODESolver.Backward') # Needed to use components [EulerImplicitSolver]  
 
     # Add gravity
     root_node.gravity = [0, -9.81, 0]  # Add gravity!
@@ -52,13 +48,13 @@ def createScene(root_node):
         rayleighMass="0.0",
         vdamping=v_damping_param,  # Damping parameter for dynamics
     )
-    solver_node.addObject("SparseLDLSolver", name="solver")
+    solver_node.addObject("SparseLDLSolver", template="CompressedRowSparseMatrixMat3x3d", name="solver")
 
     # === NEW APPROACH: Use CosseratGeometry with more sections for smoother dynamics ===
     beam_geometry_params = BeamGeometryParameters(
         beam_length=30.0,  # Same beam length
-        nb_section=30,  # 30 sections for good physics resolution
-        nb_frames=30,  # 30 frames for smooth visualization
+        nb_section=3,  # 30 sections for good physics resolution
+        nb_frames=12,  # 30 frames for smooth visualization
     )
 
     # Create geometry object
@@ -74,33 +70,21 @@ def createScene(root_node):
     base_node = _add_rigid_base(solver_node)
 
     # Create bending states with a curve (last section has more bending)
+    # We start with a slightly bent beam to better visualize the effect of gravity.
     custom_bending_states = []
     for i in range(beam_geometry.get_number_of_sections()):
-        custom_bending_states.append([0, 0.0, 0.0])
+        custom_bending_states.append([0, 0.0, 0.1])
 
     # Create cosserat state using geometry
     bending_node = _add_cosserat_state(solver_node, beam_geometry, node_name="cosserat_states",
                                        custom_bending_states=custom_bending_states)
 
     # Create cosserat frame with mass (important for dynamics!)
+    # The mass is distributed uniformly along the beam. Without mass, the beam
+    # would not be affected by gravity.
     frame_node = _add_cosserat_frame(
         base_node, bending_node, beam_geometry, beam_mass=5.0
     )
 
-
-    # === ADD FORCES ===
-    # Add a force at the tip of the beam
-    # this constance force is used only in the case we are doing force_type 1 or 2
-    force_null = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0]  # No force initially
-
-    const_force_node = frame_node.addObject('ConstantForceField', name='constForce', showArrowSize=1.e-8,
-                                                 indices=beam_geometry.nb_frames, forces=force_null)
-
-    # The effector is used only when force_type is 3
-    # create a rigid body to control the end effector of the beam
-    tip_controller = root_node.addChild('tip_controller')
-    controller_state = tip_controller.addObject('MechanicalObject', template='Rigid3d', name="controlEndEffector",
-                                                showObjectScale=0.3, position=[beam_geometry.beam_length, 0, 0, 0, 0, 0, 1],
-                                                showObject=True)
 
     return root_node
