+++ tutorials/getting_started/tutorial_04_force_interactions.py
@@ -18,11 +18,16 @@ Key improvements over manual approach:
 import os
 import sys
 
+_beam_radius = 0.5
+_beam_length = 30.
+_nb_section = 32
+
 # Add the python package to the path
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", "..", "python"))
 
 from cosserat import BeamGeometryParameters, CosseratGeometry
 
+from force_controller import ForceController
 from introduction_and_setup import (_add_cosserat_frame, _add_cosserat_state,
                                     _add_rigid_base, add_mini_header)
 
@@ -32,7 +37,10 @@ def createScene(root_node):
     """Create a Cosserat beam scene with forces and dynamics."""
     # Configure scene with time integration
     add_mini_header(root_node)
-
+    root_node.addObject('RequiredPlugin', name='Sofa.Component.LinearSolver.Direct') # Needed to use components [SparseLDLSolver]  
+    root_node.addObject('RequiredPlugin', name='Sofa.Component.ODESolver.Backward') # Needed to use components [EulerImplicitSolver]  
+    root_node.addObject('RequiredPlugin', name='Sofa.Component.MechanicalLoad') # Needed to use components [ConstantForceField] 
+    
     # Add gravity
     root_node.gravity = [0, -9.81, 0]  # Add gravity!
 
@@ -46,7 +54,7 @@ def createScene(root_node):
         rayleighMass="0.0",
         vdamping=v_damping_param,  # Damping parameter for dynamics
     )
-    solver_node.addObject("SparseLDLSolver", name="solver")
+    solver_node.addObject("SparseLDLSolver", template="CompressedRowSparseMatrixMat3x3d", name="solver")
 
     # === NEW APPROACH: Use CosseratGeometry with more sections for smoother dynamics ===
     beam_geometry_params = BeamGeometryParameters(
@@ -82,75 +90,20 @@ def createScene(root_node):
     )
 
 
-    # # -------------------------------------------------
-    # # === ADD SECOND BEAM WITH DIFFERENT PARAMETERS===
-    # # -------------------------------------------------
+    # === ADD FORCES ===
+    # Add a force at the tip of the beam
+    # this constance force is used only in the case we are doing force_type 1 or 2
+    force_null = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0]  # No force initially
 
-    # Configure time integration and solver
-    solver_node2 = root_node.addChild("solver_2")
+    const_force_node = frame_node.addObject('ConstantForceField', name='constForce', showArrowSize=1.e-8,
+                                                 indices=[beam_geometry.get_number_of_frames()], forces=force_null)
 
-    solver_node2.addObject(
-        "EulerImplicitSolver",
-        firstOrder="0",
-        rayleighStiffness="0.0",
-        rayleighMass="0.0",
-        name="euler_solver2",
-        vdamping=v_damping_param
-    )
-    solver_node2.addObject("SparseLDLSolver", name="solver2")
+    # The effector is used only when force_type is 3
+    # create a rigid body to control the end effector of the beam
+    tip_controller = root_node.addChild('tip_controller')
+    controller_state = tip_controller.addObject('MechanicalObject', template='Rigid3d', name="controlEndEffector",
+                                                showObjectScale=0.3, position=[beam_geometry.get_beam_length(), 0, 0, 0, 0, 0, 1],
+                                                showObject=True)
 
-    # # Define second beam geometry parameters
-    beam_geometry_params2 = BeamGeometryParameters(
-        beam_length=30.0,  # Same beam length
-        nb_section=15,  # 30 sections for good physics resolution
-        nb_frames=15,  # 3 frames for smooth visualization
-    )
-
-    # # Create second geometry object
-    # This beam has fewer sections (15) than the first one (30).
-    # This will make the simulation faster, but less accurate.
-    # It's a trade-off between performance and physical fidelity.
-    beam_geometry2 = CosseratGeometry(beam_geometry_params2)
-    print(f"🚀 Created second dynamic beam with:")
-    print(f"   - Length: {beam_geometry2.get_beam_length()}")
-    print(f"   - Sections: {beam_geometry2.get_number_of_sections()}")
-    print(f"   - Frames: {beam_geometry2.get_number_of_frames()}")
-    print(f"   - Mass will be distributed across frames")
-
-    # # Create rigid base for second beam
-    base_node2 = _add_rigid_base(solver_node2, node_name="rigid_base2")
-
-    # # Create cosserat state for the second beam
-    # # -------------------------------------------------
-    custom_bending_states2 = []
-    for i in range(beam_geometry2.get_number_of_sections()):
-        custom_bending_states.append([0, 0.0, 0.0])
-
-    bending_node2 = _add_cosserat_state(
-        solver_node2, beam_geometry2, node_name="cosserat_states2",
-        custom_bending_states=custom_bending_states2
-    )
-
-    # # Create cosserat frame for the second beam
-    _add_cosserat_frame(
-        base_node2, bending_node2, beam_geometry2, node_name="cosserat_in_Sofa_frame_node2",
-        beam_mass=5.0
-    )
-
-
-    # # === ADD FORCES ===
-    # # Add a force at the tip of the beam
-    # tip_frame_index = beam_geometry.get_number_of_frames()  # Last frame
-    # applied_force = [-10.0, 0.0, 0.0, 0, 0, 0]  # Force in -X direction
-    #
-    # frame_node.addObject(
-    #     "ConstantForceField",
-    #     name="tipForce",
-    #     indices=[tip_frame_index],
-    #     forces=[applied_force],
-    #     showArrowSize=1,
-    # )
-    #
-    # print(f"💪 Applied force {applied_force[:3]} at frame {tip_frame_index}")
 
     return root_node
