+++ tutorials/getting_started/tutorial_05_constraints_and_boundary_conditions.py
@@ -18,7 +18,7 @@ import sys
 # Add the python package to the path
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", "..", "python"))
 
-from python.cosserat import BeamGeometryParameters, CosseratGeometry
+from cosserat import BeamGeometryParameters, CosseratGeometry
 
 from introduction_and_setup import (_add_cosserat_frame, _add_cosserat_state,
                                     _add_rigid_base, add_mini_header)
@@ -29,6 +29,9 @@ def createScene(root_node):
     """Create a Cosserat beam scene with constraints."""
     # Configure scene with time integration
     add_mini_header(root_node)
+    root_node.addObject('RequiredPlugin', name='Sofa.Component.Constraint.Projective') # Needed to use components [FixedProjectiveConstraint]  
+    root_node.addObject('RequiredPlugin', name='Sofa.Component.LinearSolver.Direct') # Needed to use components [SparseLDLSolver]  
+    root_node.addObject('RequiredPlugin', name='Sofa.Component.ODESolver.Backward') # Needed to use components [EulerImplicitSolver]  
 
     # Add gravity
     root_node.gravity = [0, -9.81, 0]
